pyproject.toml
[project]
name = "amq-mcp-server-rabbitmq"
version = "2.2.4"
description = "A Model Context Protocol server providing access to RabbitMQ by LLMs"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Ken Liao" }]
maintainers = [{ name = "Ken Liao", email = "kenliao94@gmail.com" }]
keywords = ["rabbitmq", "mcp", "llm", "automation"]
license = { text = "Apache-2.0" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "markdownify>=0.13.1",
    "mcp>=1.6.0",
    "pika>=1.3.2",
    "protego>=0.3.1",
    "pydantic>=2.11.7",
    "readabilipy>=0.2.0",
    "requests>=2.32.3",
    "fastmcp>=2.12.0",
    "loguru>=0.7.3",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.38.0",
]

[project.scripts]
amq-mcp-server-rabbitmq = "src.server:main"

[project.optional-dependencies]
dev = ["pyright>=1.1.389", "ruff>=0.7.3", "pytest>=8.3.5"]

[project.urls]
Homepage = "https://github.com/amazon-mq/mcp-server-rabbitmq"
Source = "https://github.com/amazon-mq/mcp-server-rabbitmq"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv.workspace]
members = ["example/agent_strands"]

[tool.ruff]
line-length = 99
target-version = "py310"
exclude = ["uv.lock"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "Q"]
ignore = ["E203", "E501"]

[tool.ruff.lint.isort]
known-first-party = ["mcp_server_rabbitmq"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["src/tests"]

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B101"]

[tool.pyright]
include = ["mcp_server_rabbitmq"]
exclude = ["**/__pycache__", ".venv"]
typeCheckingMode = "basic"
pythonVersion = "3.10"
